Refactor gutter package: split renderer into focused files
Split the 1034-line ScopeLineStatusMarkerRenderer into three files
following the same separation of concerns used by IntelliJ's own
LineStatusMarkerRenderer / LineStatusMarkerPopupPanel split:

- ScopeGutterHighlighterManager: owns the RangeHighlighter, error-stripe
  marks, gutter repaint, and the anonymous LineStatusGutterMarkerRenderer
  subclass (analogous to LineStatusMarkerRenderer)
- ScopeGutterPopupPanel: builds the popup UI, toolbar actions, editor
  highlights, and rollback/copy operations (analogous to
  LineStatusMarkerPopupPanel); entirely original code, not derived from
  IntelliJ
- ScopeLineStatusMarkerRenderer: trimmed to a pure orchestrator — hover
  state, mouse listeners, filler management, and public API only
- rangeContainsY extracted as a package-level function in
  LineStatusGutterMarkerRenderer.kt so both the hover tracker and the
  gutter renderer can share it without a cross-class dependency
- NOTICE updated to reflect the new file boundaries

Also positions the gutter popup just below the last line of the clicked
range so the changed text stays visible. Prev/next navigation does not
reposition the window.
