+++ NOTICE
@@ -15,6 +15,7 @@ under the Apache License, Version 2.0:
   src/main/java/implementation/gutter/Range.kt
   src/main/java/implementation/gutter/RangesBuilder.kt
   src/main/java/implementation/gutter/ScopeDiffViewer.kt
+  src/main/java/implementation/gutter/ScopeGutterHighlighterManager.kt
   src/main/java/implementation/gutter/ScopeLineStatusMarkerRenderer.kt
 
 You may obtain a copy of the Apache License, Version 2.0 at:
